AwsVpc: Create AwsVpc class to create and connect AWS resources
Define class and method in vpc.py. The class also creates an EC2 instance
and an SSH key pair; these might be worth separating into child classes
later on.

Many of the class methods are commented out for the time being. Tests for
these methods are complicated because they can generate realresources on
AWS if not done correctly. Therefore, working with a smaller set of code
before trying to test all methods in AwsVpc.