AwsVpc: Create AwsVpc class
Define class and method in vpc.py. The class also creates an EC2 instance
and an SSH key pair; these might be worth separating into child classes
later on.

app.py just runs through all methods in the AwsVpc class. As this app
functions as expected, it will actually create live AWS resources. Be
careful when running application as of this commit.