test_vpc: Add test for aws_create_subnet()
Replaced some static value in method with arguments.

In its current form, this test only works with boto3's client interface.
The resource interface fails to stub properly and returns an error
stating that target VPC does not exist. (If using VPC ID of a real VPC,
it will actually create the resource live on AWS.)